refactor: 라우터 이름 및 스타일링 변경
